lib/vfscore: Fix eventpoll notify close
When a file is closed we have to remove all references to this
file from eventpoll instances. The current eventpoll has an open
todo, where threads would be kept sleeping even if there are no
more fds in the eventpoll. This commit adds a wakeup. It also
adds the deallocation of the fd context structure in the notify
close callback. Since we do not know which allocator has been
used for the structure, we change eventpoll to store the
allocator so that we can use it in the callback.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>