lib/vfscore: Fix list corruption in eventpoll
This commit fixes a bug in the eventpoll_mod, where the triggered
list can be corrupted if the fd is ready. The current version only adds
the fd to the triggered list, if it is already part of it. This corrupts
the list. The patch fixes this so that the fd is only added if it is NOT
already part of the list.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>